Move MultipleChoiceTaskViewModel to datacollection/tasks/multiplechoice
