+++ ground/src/main/res/layout/multiple_choice_input_task.xml
@@ -23,7 +23,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.editsubmission.MultipleChoiceTaskViewModel" />
+      type="com.google.android.ground.ui.datacollection.tasks.multiplechoice.MultipleChoiceTaskViewModel" />
   </data>
 
   <LinearLayout
